Move path-based tagging to nixpkgs
