+++ ofborg/src/tasks/eval/nixpkgs.rs
@@ -1,3 +1,6 @@
+use ofborg::tagger::{
+    MaintainerPRTagger, PathsTagger, PkgsAddedRemovedTagger, RebuildTagger, StdenvTagger,
+};
 use tasks::eval::{EvaluationStrategy, StepResult, Stdenvs, Error};
 use hubcaps::issues::IssueRef;
 use hubcaps::gists::Gists;
@@ -11,6 +14,8 @@ use ofborg::files::file_to_str;
 use ofborg::message::evaluationjob::EvaluationJob;
 use ofborg::nix::Nix;
 use ofborg::commitstatus::CommitStatus;
+use ofborg::checkout::CachedProjectCo;
+use std::collections::HashMap;
 
 pub struct NixpkgsStrategy<'a> {
     job: &'a EvaluationJob,
@@ -18,16 +23,18 @@ pub struct NixpkgsStrategy<'a> {
     gists: &'a Gists<'a>,
     events: (),
     nix: &'a Nix,
+    tag_paths: &'a HashMap<String, Vec<String>>,
 }
 
 impl <'a> NixpkgsStrategy<'a> {
-    pub fn new(job: &'a EvaluationJob, issue: &'a IssueRef, gists: &'a Gists, nix: &'a Nix) -> NixpkgsStrategy<'a> {
+    pub fn new(job: &'a EvaluationJob, issue: &'a IssueRef, gists: &'a Gists, nix: &'a Nix, tag_paths: &'a HashMap<String, Vec<String>>,) -> NixpkgsStrategy<'a> {
         Self {
             job,
             issue,
             gists,
             nix,
-            events: ()
+            tag_paths,
+            events: (),
         }
     }
 
@@ -42,6 +49,16 @@ impl <'a> NixpkgsStrategy<'a> {
         if darwin {
             update_labels(&self.issue, &[String::from("6.topic: darwin")], &[]);
         }
+    }
+
+    fn tag_from_paths(&self, issue: &hubcaps::issues::IssueRef, paths: &[String]) {
+        let mut tagger = PathsTagger::new(self.tag_paths.clone());
+
+        for path in paths {
+            tagger.path_changed(&path);
+        }
+
+        update_labels(&issue, &tagger.tags_to_add(), &tagger.tags_to_remove());
     }
 }
 
@@ -104,5 +121,84 @@ impl <'a> EvaluationStrategy for NixpkgsStrategy<'a> {
             .notify(Event::EvaluationDurationCount(target_branch.clone()));
 */
         Ok(())
+    }
+
+    fn after_fetch(&self, co: &CachedProjectCo) -> StepResult {
+        let possibly_touched_packages = parse_commit_messages(
+            &co.commit_messages_from_head(&self.job.pr.head_sha)
+                .unwrap_or_else(|_| vec!["".to_owned()]),
+        );
+
+        let changed_paths = co
+            .files_changed_from_head(&self.job.pr.head_sha)
+            .unwrap_or_else(|_| vec![]);
+        self.tag_from_paths(&self.issue, &changed_paths);
+
+        Ok(())
+    }
+}
+
+
+fn parse_commit_messages(messages: &[String]) -> Vec<String> {
+    messages
+        .iter()
+        .filter_map(|line| {
+            // Convert "foo: some notes" in to "foo"
+            let parts: Vec<&str> = line.splitn(2, ':').collect();
+            if parts.len() == 2 {
+                Some(parts[0])
+            } else {
+                None
+            }
+        })
+        .flat_map(|line| {
+            let pkgs: Vec<&str> = line.split(',').collect();
+            pkgs
+        })
+        .map(|line| line.trim().to_owned())
+        .collect()
+}
+
+
+#[cfg(test)]
+mod tests {
+
+    use super::*;
+
+    #[test]
+    fn test_parse_commit_messages() {
+        let expect: Vec<&str> = vec![
+            "firefox{-esr", // don't support such fancy syntax
+            "}",            // Don't support such fancy syntax
+            "firefox",
+            "buildkite-agent",
+            "python.pkgs.ptyprocess",
+            "python.pkgs.ptyprocess",
+            "android-studio-preview",
+            "foo",
+            "bar",
+        ];
+        assert_eq!(
+            parse_commit_messages(
+                &"
+              firefox{-esr,}: fix failing build due to the google-api-key
+              Merge pull request #34483 from andir/dovecot-cve-2017-15132
+              firefox: enable official branding
+              Merge pull request #34442 from rnhmjoj/virtual
+              buildkite-agent: enable building on darwin
+              python.pkgs.ptyprocess: 0.5 -> 0.5.2
+              python.pkgs.ptyprocess: move expression
+              Merge pull request #34465 from steveeJ/steveej-attempt-qtile-bump-0.10.7
+              android-studio-preview: 3.1.0.8 -> 3.1.0.9
+              Merge pull request #34188 from dotlambda/home-assistant
+              Merge pull request #34414 from dotlambda/postfix
+              foo,bar: something here: yeah
+            "
+                .lines()
+                .map(|l| l.to_owned())
+                .collect::<Vec<String>>(),
+            ),
+            expect
+        );
     }
 }
